AC-38:  Refactor presentation
