+++ presentation/ui/src/test/java/com/jfmr/ac/test/presentation/ui/character/detail/viewmodel/DetailViewModelTest.kt
@@ -10,9 +10,9 @@ import com.jfmr.ac.test.presentation.ui.character.list.model.toUI
 import com.jfmr.ac.test.presentation.ui.utils.CharacterUIUtils
 import com.jfmr.ac.test.tests.MainCoroutineRule
 import com.jfmr.ac.test.tests.character.CharacterUtils
-import com.jfmr.ac.test.usecase.character.CharacterDetailUseCase
-import com.jfmr.ac.test.usecase.character.UpdateCharacterUseCase
-import com.jfmr.ac.test.usecase.episode.GetEpisodesUseCase
+import com.jfmr.ac.test.usecase.character.detail.CharacterDetailUseCase
+import com.jfmr.ac.test.usecase.character.update.UpdateCharacterUseCase
+import com.jfmr.ac.test.usecase.episode.list.GetEpisodesUseCase
 import io.mockk.MockKAnnotations
 import io.mockk.clearAllMocks
 import io.mockk.coEvery
