+++ test/src/main/java/com/jfmr/ac/test/tests/episodes/EpisodeUtils.kt
@@ -1,6 +1,6 @@
 package com.jfmr.ac.test.tests.episodes
 
-import com.jfmr.ac.test.data.api.rickandmorty.episode.entity.EpisodeResponse
+import com.jfmr.ac.test.data.api.rickandmorty.dto.episode.entity.EpisodeResponse
 import com.jfmr.ac.test.domain.model.episode.Episode
 import com.jfmr.ac.test.tests.TestUtils
